+++ src/GameCollector.StoreHandlers.Amazon/ExternalWebsites.cs
@@ -1,0 +1,40 @@
+namespace GameCollector.StoreHandlers.Amazon;
+
+internal record ExternalWebsites
+{
+    public string? Youtube { get; init; }
+
+    public string? Twitter { get; init; }
+
+    public string? Twitch { get; init; }
+
+    public string? Facebook { get; init; }
+
+    public string? Reddit { get; init; }
+
+    public string? Official { get; init; }
+
+    public string? Steam { get; init; }
+
+    public string? Epicgames { get; init; }
+
+    public string? Instagram { get; init; }
+
+    public string? Wikipedia { get; init; }
+
+    public string? Support { get; init; }
+
+    public string? Igdb { get; init; }
+
+    public string? Gog { get; init; }
+
+    public string? Itch { get; init; }
+
+    public string? Wikia { get; init; }
+
+    public string? Android { get; init; }
+
+    public string? Ipad { get; init; }
+
+    public string? Iphone { get; init; }
+}