+++ src/GameCollector.StoreHandlers.Amazon/SourceGenerationContext.cs
@@ -1,0 +1,7 @@
+using System.Text.Json.Serialization;
+
+namespace GameCollector.StoreHandlers.Amazon;
+
+[JsonSourceGenerationOptions(WriteIndented = false, GenerationMode = JsonSourceGenerationMode.Default)]
+[JsonSerializable(typeof(ExternalWebsites))]
+internal partial class SourceGenerationContext : JsonSerializerContext { }